% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_integrity.R
\name{check_data_integrity}
\alias{check_data_integrity}
\title{check_data_integrity}
\usage{
check_data_integrity(dataset)
}
\arguments{
\item{dataset}{a dataset}
}
\value{
A message
}
\description{
check_data_integrity
}
\examples{


# Example dataset

datafile <- system.file("nyc_squirrels_sample.csv", package = "integrity")
nyc_squirrels <- read.csv(datafile)

head(nyc_squirrels)
  
check_data_integrity(dataset = nyc_squirrels)
}
